requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-binance==1.0.17
web3==6.5.0
sqlalchemy==2.0.17
pydantic==2.0.2
python-dotenv==1.0.0